#134 [feat] 데이터 업로드 서비스 메서드를 구현한다.

+++ src/main/java/com/dataracy/modules/data/adapter/persistence/entity/DataEntity.java
@@ -53,9 +53,9 @@ public class DataEntity extends BaseEntity {
 
     // 캐싱 필드
     @Column(nullable = false)
-    private Long downloadCount;
+    private int downloadCount;
     @Column(nullable = false)
-    private Long recentWeekDownloadCount;
+    private int recentWeekDownloadCount;
 
     // 메타데이터 FK (1:1)
     @OneToOne(mappedBy = "data", cascade = CascadeType.PERSIST)
@@ -90,8 +90,8 @@ public class DataEntity extends BaseEntity {
             String analysisGuide,
             String dataFileUrl,
             String thumbnailUrl,
-            Long downloadCount,
-            Long recentWeekDownloadCount,
+            int downloadCount,
+            int recentWeekDownloadCount,
             DataMetadataEntity metadata
     ) {
         return DataEntity.builder()
